Including commit to change the layout

+++ src/EProductivity.Core/Model/Area.cs
@@ -4,7 +4,7 @@ namespace EProductivity.Core.Model
 {
     public class Area
     {
-        public int AreaId { get; set; }
+        public long AreaId { get; set; }
         public string Name { get; set; }
         public Organization Organization { get; set; }
         public long OrganizationId { get; set; }
